Fixed error message, it seems to work for rain now. More testing needed, though
