+++ src/test/java/org/terasology/weatherManager/WeatherConditionTest.java
@@ -17,7 +17,6 @@ package org.terasology.weatherManager;
 
 import org.junit.Test;
 import org.terasology.weatherManager.systems.MarkovChainWeatherGenerator;
-import org.terasology.world.time.WorldTime;
 
 /**
  * Created by Linus on 5-11-2014.
